+++ tests/snowpark/test_snowpark_entity.py
@@ -3,6 +3,7 @@ from unittest import mock
 
 import pytest
 import yaml
+from click import UsageError
 from snowflake.cli._plugins.snowpark.package.anaconda_packages import (
     AnacondaPackages,
     AvailablePackage,
@@ -229,7 +230,21 @@ def test_get_usage_grant_sql(example_function_workspace):
 def test_get_deploy_sql_with_repository_packages(example_function_workspace, snapshot):
     entity, _ = example_function_workspace
     entity.model.artifact_repository = "snowflake.snowpark.pypi_shared_repository"
-    entity.model.artifact_repository_packages = ["package1", "package2"]
+    entity.model.packages = ["package1", "package2"]
     entity.model.resource_constraint = {"architecture": "x86"}
     deploy_sql = entity.get_deploy_sql(CreateMode.create)
     assert deploy_sql == snapshot
+
+
+def test_get_deploy_sql_with_packages_and_artifact_repository_packages(
+    example_function_workspace,
+):
+    entity, _ = example_function_workspace
+    entity.model.artifact_repository = "snowflake.snowpark.pypi_shared_repository"
+    entity.model.packages = ["package1", "package2"]
+    entity.model.artifact_repository_packages = ["package1", "package2"]
+    with pytest.raises(
+        UsageError,
+        match="You cannot specify both artifact_repository_packages and packages.",
+    ):
+        entity.get_deploy_sql(CreateMode.create)